Modify the distanceDampener function
DistanceDampener previously squared the proximity lerp, so that when EE was proximal to target the limb would quickly stop moving.

Dampener is now root of proximity lerp, so limb actually continues to move until extremely close to target.

This encourages EE to move much closer to target before motion halts